restructuring client
