+++ client/Cargo.toml
@@ -72,9 +72,6 @@ path = "../substratee-worker-primitives"
 [dependencies.substratee-stf]
 path = "../stf"
 
-[dependencies.substratee-client-primitives]
-path = "../substratee-client-primitives"
-
 [dependencies.substratee-worker-api]
 path = "../worker/worker-api"
