--- substratee-client-primitives/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "substratee-client-primitives"
-version = "0.8.0"
-authors = ["Supercomputing Systems AG <info@scs.ch>"]
-edition = "2018"
-
-[dependencies]
-log = "0.4"
-env_logger = "0.7"
-hex = "0.4.2"
-json = "0.12.0"
-substrate-bip39 = "0.4.2"
-tiny-bip39 = "0.6.2"
-serde_json = "1.0"
-clap = "2.33"
-clap-nested = "0.3.1"
-primitive-types = { version = "0.6.1", default-features = false, features = ["codec"] }
-base58 			        = "0.1"
-chrono = "*"
-blake2-rfc      = { version = "0.2.18", default-features = false}
-geojson = "0.17"
-sgx_crypto_helper = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
-
-[dependencies.serde]
-features = ["derive"]
-version = "1.0"
-
-[dependencies.codec]
-default-features = false
-package = "parity-scale-codec"
-features = ["derive"]
-version = "2.0.0"
