--- substratee-client-primitives/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod common_args;
-pub mod common_args_processing;
